fix/page change
