+++ src/main.js
@@ -3,6 +3,7 @@ import './assets/main.css'
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 import App from './App.vue'
+import router from "./router";
 
 const app = createApp(App)
 app.use(createPinia())
